+++ app/src/main/java/org/sopt/at/data/dto/request/NicknameChangeRequestDto.kt
