+++ app/src/main/java/org/sopt/at/data/dto/request/SignInRequestDto.kt
@@ -1,13 +1,12 @@
 package org.sopt.at.data.dto.request
+
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
 
 @Serializable
-data class SignUpRequest(
+data class SignInRequestDto(
     @SerialName("loginId")
     val loginId: String,
     @SerialName("password")
     val password: String,
-    @SerialName("nickname")
-    val nickname: String
 )
