+++ app/src/main/java/org/sopt/at/data/dto/request/SignUpRequestDto.kt
@@ -1,12 +1,13 @@
 package org.sopt.at.data.dto.request
-
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
 
 @Serializable
-data class SignInRequest(
+data class SignUpRequestDto(
     @SerialName("loginId")
     val loginId: String,
     @SerialName("password")
     val password: String,
+    @SerialName("nickname")
+    val nickname: String
 )
