+++ app/src/main/java/org/sopt/at/data/dto/response/SignUpResponseDto.kt
@@ -4,8 +4,9 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
 
 @Serializable
-data class UserInfoResponseDto(
+data class SignUpResponseDto(
+    @SerialName("userId")
+    val userId: Int,
     @SerialName("nickname")
     val nickname: String
 )
-
