+++ app/src/main/java/org/sopt/at/data/dto/response/UserInfoResponseDto.kt
@@ -3,11 +3,8 @@ package org.sopt.at.data.dto.response
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
 
-
 @Serializable
-data class SignUpResponseDto(
-    @SerialName("userId")
-    val userId: Int,
+data class UserInfoResponseDto(
     @SerialName("nickname")
     val nickname: String
 )
