+++ app/src/main/java/org/sopt/at/ui/main/MainActivity.kt
@@ -24,18 +24,3 @@ class MainActivity : ComponentActivity() {
     }
 }
 
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    ATSOPTANDROIDTheme {
-        Greeting("Android")
-    }
-}
